requirements.txt
tensorflow-gpu==1.0.0
jieba
